howard revision of heatmap code
todo: 1. add in multi-participant heatmap, and 2. make sure the x and y axes plotted are in the correct coordinate system, it should be pretty clear in the search images that the first fixation tends to be on the red turtle. 3. maybe adjust heatmaps so the important turtles are labeled in some way so you can see them through the heatmap